+++ uikit/src/main/java/com/sendbird/uikit/internal/ui/viewholders/NotificationTimelineViewHolder.kt
@@ -1,6 +1,6 @@
 package com.sendbird.uikit.internal.ui.viewholders
 
-import com.sendbird.android.channel.BaseChannel
+import com.sendbird.android.channel.GroupChannel
 import com.sendbird.android.message.BaseMessage
 import com.sendbird.uikit.databinding.SbViewTimeLineMessageBinding
 import com.sendbird.uikit.internal.model.notifications.NotificationConfig
@@ -9,7 +9,7 @@ internal class NotificationTimelineViewHolder internal constructor(
     val binding: SbViewTimeLineMessageBinding,
 ) : NotificationViewHolder(binding.root) {
 
-    override fun bind(channel: BaseChannel, message: BaseMessage, uiConfig: NotificationConfig?) {
-        binding.timelineMessageView.drawTimeline(message, uiConfig)
+    override fun bind(channel: GroupChannel, message: BaseMessage, config: NotificationConfig?) {
+        binding.timelineMessageView.drawTimeline(message, config)
     }
 }